# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WTD Lead Generation API",
    "description": "Workspace routes require X-API-Key. Microsoft login is not required for research or editing. Approving, sending and resolving email require a connected Microsoft mailbox via X-Session-Token.",
    "version": "3.1.0"
  },
  "paths": {
...
    "/api/v1/auth/microsoft/start": {
      "post": {
        "tags": [
          "Microsoft mailbox connection"
        ],
        "summary": "Microsoft Start",
        "operationId": "microsoft_start_api_v1_auth_microsoft_start_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BackendKey": []
          }
        ]
      }
    },
    "/api/v1/auth/microsoft/callback": {
      "post": {
        "tags": [
          "Microsoft mailbox connection"
        ],
        "summary": "Microsoft Callback",
        "operationId": "microsoft_callback_api_v1_auth_microsoft_callback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MicrosoftCallback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BackendKey": []
          }
        ]
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "tags": [
          "Microsoft mailbox connection"
        ],
        "summary": "Session Info",
        "operationId": "session_info_api_v1_auth_session_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BackendKey": []
          },
          {
            "MicrosoftMailbox": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Microsoft mailbox connection"
        ],
        "summary": "Logout",
        "operationId": "logout_api_v1_auth_session_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BackendKey": []
          },
          {
            "MicrosoftMailbox": []
          }
        ]
      }
    },
    "/api/v1/settings": {
      "get": {
        "tags": [
          "Configuration"
        ],
        "summary": "Public Settings",
        "operationId": "public_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BackendKey": []
          }
        ]
      }
    },
    "/api/v1/runs": {
      "post": {
        "tags": [
          "Searches"
        ],
        "summary": "Create Run",
        "operationId": "create_run_api_v1_runs_post",
        "security": [
          {
            "BackendKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Searches"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_v1_runs_get",
        "security": [
          {
            "BackendKey": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "Searches"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "security": [
          {
            "BackendKey": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/export": {
      "get": {
        "tags": [
          "Searches"
        ],
        "summary": "Export",
        "operationId": "export_api_v1_runs__run_id__export_get",
        "security": [
          {
            "BackendKey": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "csv",
                "json"
              ],
              "type": "string",
              "default": "csv",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads": {
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "List Leads",
        "operationId": "list_leads_api_v1_leads_get",
        "security": [
          {
            "BackendKey": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "name": "tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "Cool",
                    "Warm",
                    "Hot"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tier"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads/{lead_id}": {
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "Get Lead",
        "operationId": "get_lead_api_v1_leads__lead_id__get",
        "security": [
          {
            "BackendKey": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/emails": {
      "get": {
        "tags": [
          "Email review"
        ],
        "summary": "List Emails",
        "operationId": "list_emails_api_v1_emails_get",
        "security": [
          {
            "BackendKey": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "draft",
                    "ready",
                    "sending",
                    "sent",
                    "failed",
                    "unknown"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/emails/{email_id}": {
      "get": {
        "tags": [
          "Email review"
        ],
        "summary": "Get Email",
        "operationId": "get_email_api_v1_emails__email_id__get",
        "security": [
          {
            "BackendKey": []
          }
        ],
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Email review"
        ],
        "summary": "Edit Email",
        "operationId": "edit_email_api_v1_emails__email_id__patch",
        "security": [
          {
            "BackendKey": []
          }
        ],
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/emails/{email_id}/approval": {
      "post": {
        "tags": [
          "Email review"
        ],
        "summary": "Approve Email",
        "operationId": "approve_email_api_v1_emails__email_id__approval_post",
        "security": [
          {
            "BackendKey": []
          }
        ],
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Approval"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "description": "Setting ready_to_send=true also requires X-Session-Token for the sending mailbox. Removing approval requires only X-API-Key."
      }
    },
...
    "/api/v1/zoominfo/test-connection": {
      "post": {
        "tags": [
          "ZoomInfo"
        ],
        "summary": "Test Zoominfo Connection",
        "description": "Check OAuth and a fresh topic lookup; does not enrich contacts or send email.",
        "operationId": "test_zoominfo_connection_api_v1_zoominfo_test_connection_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BackendKey": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MicrosoftCallback": {
        "properties": {
          "flow_handle": {
            "type": "string",
            "maxLength": 43,
            "minLength": 43,
            "title": "Flow Handle"
          },
          "response": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Response"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "flow_handle",
          "response"
        ],
        "title": "MicrosoftCallback"
      },
...
    },
    "securitySchemes": {
      "BackendKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "MicrosoftMailbox": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Token"
      }
    }
  }
}